+++ src/integration-test/java/org/openlmis/notification/web/WireMockResponses.java
@@ -17,14 +17,16 @@ package org.openlmis.notification.web;
 
 interface WireMockResponses {
 
-  String MOCK_CHECK_RESULT = "{"
-      + "  \"aud\": [\n"
-      + "    \"auth\",\n"
-      + "    \"example\",\n"
-      + "    \"requisition\",\n"
-      + "    \"notification\",\n"
-      + "    \"referencedata\"\n"
-      + "  ],\n"
+  String MOCK_TOKEN_REQUEST_RESPONSE = "{"
+      + "  \"access_token\": \"418c89c5-7f21-4cd1-a63a-38c47892b0fe\",\n"
+      + "  \"token_type\": \"bearer\",\n"
+      + "  \"expires_in\": 847,\n"
+      + "  \"scope\": \"read write\",\n"
+      + "  \"referenceDataUserId\": \"35316636-6264-6331-2d34-3933322d3462\"\n"
+      + "}";
+
+  String MOCK_USER_CHECK_RESULT = "{\n"
+      + "  \"aud\": [\"auth\",\"example\",\"requisition\",\"notification\",\"referencedata\"],\n"
       + "  \"user_name\": \"admin\",\n"
       + "  \"referenceDataUserId\": \"35316636-6264-6331-2d34-3933322d3462\",\n"
       + "  \"scope\": [\n"
@@ -39,11 +41,12 @@ interface WireMockResponses {
       + "  \"client_id\": \"trusted-client\"\n"
       + "}";
 
-  String MOCK_TOKEN_REQUEST_RESPONSE = "{"
-      + "  \"access_token\": \"418c89c5-7f21-4cd1-a63a-38c47892b0fe\",\n"
-      + "  \"token_type\": \"bearer\",\n"
-      + "  \"expires_in\": 847,\n"
-      + "  \"scope\": \"read write\",\n"
-      + "  \"referenceDataUserId\": \"35316636-6264-6331-2d34-3933322d3462\"\n"
+  String MOCK_SERVICE_CHECK_RESULT = "{\n"
+      + "  \"aud\": [\"auth\",\"example\",\"requisition\",\"notification\",\"referencedata\"],\n"
+      + "  \"scope\": [\"read\", \"write\"],\n"
+      + "  \"exp\": 1474500343,\n"
+      + "  \"authorities\": [\"TRUSTED_CLIENT\"],\n"
+      + "  \"client_id\": \"trusted-client\"\n"
       + "}";
+
 }